helper/web/requirements.txt
Flask==2.0.3
Flask_Cors==3.0.10
Flask_RESTful==0.3.9
pycryptodome==3.15.0
PyMySQL==1.0.2
Werkzeug==2.0.3
cryptography
paddlepaddle
paddleocr>=2.0.1
